Reorganize utils
- Move all functions out of __init__ and into appropriate submodule
- Move `approx_equals` into test/_utils.py (since we only use it for
  tests)
- Rename test/util to test/utils

BREAKING CHANGE: (for users)
- `gpytorch.utils.scale_to_bounds` is now
  `gpytorch.utils.grid.scale_to_bounds`

Important for developers:
- `gpytorch.utils.approx_equals` is moved into a test-specific folder
