+++ gpytorch/means/constant_mean.py
@@ -4,7 +4,7 @@ from __future__ import print_function
 from __future__ import unicode_literals
 
 import torch
-from gpytorch.means import Mean
+from .mean import Mean
 
 
 class ConstantMean(Mean):
