+++ gpytorch/utils/__init__.py
@@ -3,175 +3,16 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-import torch
-from copy import deepcopy
-from operator import mul
-from .interpolation import Interpolation
 from .linear_cg import linear_cg
-from . import grid
-from . import pivoted_cholesky
-from . import lanczos
-from . import sparse
-from .grid import scale_to_bounds
-from .eig import batch_symeig
-from .cholesky import batch_potrf, batch_potrs
 from .stochastic_lq import StochasticLQ
-
-
-def approx_equal(self, other, epsilon=1e-4):
-    """
-    Determines if two tensors are approximately equal
-    Args:
-        - self: tensor
-        - other: tensor
-    Returns:
-        - bool
-    """
-    if self.size() != other.size():
-        raise RuntimeError(
-            "Size mismatch between self ({self}) and other ({other})".format(self=self.size(), other=other.size())
-        )
-    return torch.max((self - other).abs()) <= epsilon
-
-
-def bdsmm(sparse, dense):
-    """
-    Batch dense-sparse matrix multiply
-    """
-    if sparse.ndimension() > 2:
-        batch_size, n_rows, n_cols = sparse.size()
-        batch_assignment = sparse._indices()[0]
-        indices = sparse._indices()[1:].clone()
-        indices[0].add_(n_rows, batch_assignment)
-        indices[1].add_(n_cols, batch_assignment)
-        sparse_2d = torch.sparse_coo_tensor(
-            indices,
-            sparse._values(),
-            torch.Size((batch_size * n_rows, batch_size * n_cols)),
-            dtype=sparse._values().dtype,
-            device=sparse._values().device,
-        )
-
-        if dense.size(0) == 1:
-            dense = dense.repeat(batch_size, 1, 1)
-        dense_2d = dense.contiguous().view(batch_size * n_cols, -1)
-        res = torch.dsmm(sparse_2d, dense_2d)
-        res = res.view(batch_size, n_rows, -1)
-        return res
-    elif dense.ndimension() == 3:
-        batch_size, _, n_cols = dense.size()
-        res = torch.dsmm(sparse, dense.transpose(0, 1).contiguous().view(-1, batch_size * n_cols))
-        res = res.view(-1, batch_size, n_cols)
-        res = res.transpose(0, 1).contiguous()
-        return res
-    else:
-        return torch.dsmm(sparse, dense)
-
-
-def left_interp(interp_indices, interp_values, rhs):
-    """
-    """
-    is_vector = rhs.ndimension() == 1
-
-    if is_vector:
-        res = rhs.index_select(0, interp_indices.view(-1)).view(*interp_values.size())
-        res = res.mul(interp_values)
-        res = res.sum(-1)
-        return res
-
-    else:
-        # Special cuda version -- this is faster on the GPU for some reason
-        if interp_indices.is_cuda:
-            if interp_indices.ndimension() == 3:
-                n_batch, n_data, n_interp = interp_indices.size()
-                interp_indices = interp_indices.contiguous().view(-1)
-                interp_values = interp_values.contiguous().view(-1, 1)
-
-                if rhs.ndimension() == 3:
-                    if rhs.size(0) == 1 and interp_indices.size(0) > 1:
-                        rhs = rhs.expand(interp_indices.size(0), rhs.size(1), rhs.size(2))
-                    batch_indices = torch.arange(0, n_batch, dtype=torch.long, device=rhs.device).unsqueeze_(1)
-                    batch_indices = batch_indices.repeat(1, n_data * n_interp).view(-1)
-                    res = rhs[batch_indices, interp_indices, :] * interp_values
-                else:
-                    res = rhs[interp_indices, :].unsqueeze(0) * interp_values
-                res = res.view(n_batch, n_data, n_interp, -1)
-                res = res.sum(-2)
-                return res
-            else:
-                n_data, n_interp = interp_indices.size()
-                interp_indices = interp_indices.contiguous().view(-1)
-                interp_values = interp_values.contiguous().view(-1, 1)
-                if rhs.ndimension() == 3:
-                    n_batch, _, n_cols = rhs.size()
-                    rhs = rhs.transpose(0, 1).contiguous().view(-1, n_batch * n_cols)
-                    res = rhs[interp_indices, :] * interp_values
-                    res = res.view(n_data, n_interp, n_batch, n_cols)
-                    res = res.sum(-2).transpose(0, 1).contiguous()
-                else:
-                    res = rhs[interp_indices, :] * interp_values
-                    res = res.view(n_data, n_interp, -1)
-                    res = res.sum(-2)
-                return res
-
-        # Special non-cuda version -- this is faster on the CPU
-        else:
-            interp_size = list(interp_indices.size()) + [rhs.size(-1)]
-            rhs_size = deepcopy(interp_size)
-            rhs_size[-3] = rhs.size()[-2]
-            interp_indices_expanded = interp_indices.unsqueeze(-1).expand(*interp_size)
-            res = rhs.unsqueeze(-2).expand(*rhs_size).gather(-3, interp_indices_expanded)
-            res = res.mul(interp_values.unsqueeze(-1).expand(interp_size))
-            return res.sum(-2)
-
-
-def left_t_interp(interp_indices, interp_values, rhs, output_dim):
-    """
-    """
-    from .. import dsmm
-
-    is_vector = rhs.ndimension() == 1
-    if is_vector:
-        rhs = rhs.unsqueeze(-1)
-
-    is_batch = rhs.ndimension() == 3
-    if not is_batch:
-        rhs = rhs.unsqueeze(0)
-    if not interp_indices.ndimension() == 3:
-        interp_indices = interp_indices.unsqueeze(0)
-        interp_values = interp_values.unsqueeze(0)
-
-    batch_size, n_data, n_interp = interp_values.size()
-    _, _, n_cols = rhs.size()
-
-    values = (rhs.unsqueeze(-2) * interp_values.unsqueeze(-1)).view(batch_size, n_data * n_interp, n_cols)
-
-    flat_interp_indices = interp_indices.contiguous().view(1, -1)
-    batch_indices = torch.arange(0, batch_size, dtype=torch.long, device=values.device).unsqueeze_(1)
-    batch_indices = batch_indices.repeat(1, n_data * n_interp).view(1, -1)
-    column_indices = torch.arange(0, n_data * n_interp, dtype=torch.long, device=values.device).unsqueeze_(1)
-    column_indices = column_indices.repeat(batch_size, 1).view(1, -1)
-
-    summing_matrix_indices = torch.cat([batch_indices, flat_interp_indices, column_indices])
-    summing_matrix_values = torch.ones(
-        batch_size * n_data * n_interp, dtype=interp_values.dtype, device=interp_values.device
-    )
-    size = torch.Size((batch_size, output_dim, n_data * n_interp))
-
-    type_name = summing_matrix_values.type().split(".")[-1]  # e.g. FloatTensor
-    if interp_values.is_cuda:
-        cls = getattr(torch.cuda.sparse, type_name)
-    else:
-        cls = getattr(torch.sparse, type_name)
-    summing_matrix = cls(summing_matrix_indices, summing_matrix_values, size)
-
-    res = dsmm(summing_matrix, values)
-
-    if not is_batch:
-        res = res.squeeze(0)
-    if is_vector:
-        res = res.squeeze(-1)
-    return res
+from . import cholesky
+from . import eig
+from . import fft
+from . import grid
+from . import interpolation
+from . import lanczos
+from . import pivoted_cholesky
+from . import sparse
 
 
 def prod(items):
@@ -186,234 +27,15 @@ def prod(items):
         return 1
 
 
-def sparse_eye(size):
-    """
-    Returns the identity matrix as a sparse matrix
-    """
-    indices = torch.arange(0, size).long().unsqueeze(0).expand(2, size)
-    values = torch.tensor(1.).expand(size)
-    cls = getattr(torch.sparse, values.type().split(".")[-1])
-    return cls(indices, values, torch.Size([size, size]))
-
-
-def sparse_getitem(sparse, idxs):
-    """
-    """
-    if not isinstance(idxs, tuple):
-        idxs = (idxs,)
-
-    if not sparse.ndimension() <= 2:
-        raise RuntimeError("Must be a 1d or 2d sparse tensor")
-
-    if len(idxs) > sparse.ndimension():
-        raise RuntimeError("Invalid index for %d-order tensor" % sparse.ndimension())
-
-    indices = sparse._indices()
-    values = sparse._values()
-    size = list(sparse.size())
-
-    for i, idx in list(enumerate(idxs))[::-1]:
-        if isinstance(idx, int):
-            del size[i]
-            mask = indices[i].eq(idx)
-            if sum(mask):
-                new_indices = torch.zeros(indices.size(0) - 1, sum(mask), dtype=indices.dtype, device=indices.device)
-                for j in range(indices.size(0)):
-                    if i > j:
-                        new_indices[j].copy_(indices[j][mask])
-                    elif i < j:
-                        new_indices[j - 1].copy_(indices[j][mask])
-                indices = new_indices
-                values = values[mask]
-            else:
-                indices.resize_(indices.size(0) - 1, 1).zero_()
-                values.resize_(1).zero_()
-
-            if not len(size):
-                return sum(values)
-
-        elif isinstance(idx, slice):
-            start, stop, step = idx.indices(size[i])
-            size = list(size[:i]) + [stop - start] + list(size[i + 1 :])
-            if step != 1:
-                raise RuntimeError("Slicing with step is not supported")
-            mask = indices[i].lt(stop) * indices[i].ge(start)
-            if sum(mask):
-                new_indices = torch.zeros(indices.size(0), sum(mask), dtype=indices.dtype, device=indices.device)
-                for j in range(indices.size(0)):
-                    new_indices[j].copy_(indices[j][mask])
-                new_indices[i].sub_(start)
-                indices = new_indices
-                values = values[mask]
-            else:
-                indices.resize_(indices.size(0), 1).zero_()
-                values.resize_(1).zero_()
-
-        else:
-            raise RuntimeError("Unknown index type")
-
-    return torch.sparse_coo_tensor(indices, values, torch.Size(size), dtype=values.dtype, device=values.device)
-
-
-def sparse_repeat(sparse, *repeat_sizes):
-    """
-    """
-    orig_ndim = sparse.ndimension()
-    new_ndim = len(repeat_sizes)
-    orig_nvalues = sparse._indices().size(1)
-
-    # Expand the number of dimensions to match repeat_sizes
-    indices = torch.cat(
-        [
-            torch.zeros(new_ndim - orig_ndim, orig_nvalues, dtype=torch.long, device=sparse._indices().device),
-            sparse._indices(),
-        ]
-    )
-    values = sparse._values()
-    size = [1] * (new_ndim - orig_ndim) + list(sparse.size())
-
-    # Expand each dimension
-    new_indices = torch.zeros(
-        indices.size(0), indices.size(1) * mul(*repeat_sizes), dtype=indices.dtype, device=indices.device
-    )
-    new_values = values.repeat(mul(*repeat_sizes))
-    new_size = [dim_size * repeat_size for dim_size, repeat_size in zip(size, repeat_sizes)]
-
-    # Fill in new indices
-    new_indices[:, :orig_nvalues].copy_(indices)
-    unit_size = orig_nvalues
-    for i in range(new_ndim)[::-1]:
-        repeat_size = repeat_sizes[i]
-        for j in range(1, repeat_size):
-            new_indices[:, unit_size * j : unit_size * (j + 1)].copy_(new_indices[:, :unit_size])
-            new_indices[i, unit_size * j : unit_size * (j + 1)] += j * size[i]
-        unit_size *= repeat_size
-
-    return torch.sparse_coo_tensor(
-        new_indices, new_values, torch.Size(new_size), dtype=new_values.dtype, device=new_values.device
-    )
-
-
-def to_sparse(dense):
-    """
-    """
-    mask = dense.ne(0)
-    indices = mask.nonzero()
-    if indices.storage():
-        values = dense[mask]
-    else:
-        indices = indices.resize_(1, dense.ndimension()).zero_()
-        values = torch.tensor(0, dtype=dense.dtype, device=dense.device)
-
-    # Construct sparse tensor
-    klass = getattr(torch.sparse, dense.type().split(".")[-1])
-    res = klass(indices.t(), values, dense.size())
-    if dense.is_cuda:
-        res = res.cuda()
-    return res
-
-
-def tridiag_batch_potrf(trid, upper=False):
-    """
-    """
-    if not torch.is_tensor(trid):
-        raise RuntimeError("tridiag_batch_potrf is only defined for tensors")
-
-    batch_size, diag_size, _ = trid.size()
-    batch_index = torch.arange(0, batch_size, dtype=torch.long, device=trid.device)
-    off_batch_index = batch_index.unsqueeze(1).repeat(diag_size - 1, 1).view(-1)
-    batch_index = batch_index.unsqueeze(1).repeat(diag_size, 1).view(-1)
-    diag_index = torch.arange(0, diag_size, dtype=torch.long, device=trid.device)
-    diag_index = diag_index.unsqueeze(1).repeat(1, batch_size).view(-1)
-    off_diag_index = torch.arange(0, diag_size - 1, dtype=torch.long, device=trid.device)
-    off_diag_index = off_diag_index.unsqueeze(1).repeat(1, batch_size).view(-1)
-
-    t_main_diag = trid[batch_index, diag_index, diag_index].view(diag_size, batch_size)
-    t_off_diag = trid[off_batch_index, off_diag_index + 1, off_diag_index].view(diag_size - 1, batch_size)
-
-    chol_main_diag = torch.empty_like(t_main_diag)
-    chol_off_diag = torch.empty_like(t_off_diag)
-
-    chol_main_diag[0].copy_(t_main_diag[0].sqrt())
-    for i in range(1, diag_size):
-        chol_off_diag[i - 1].copy_(t_off_diag[i - 1] / chol_main_diag[i - 1])
-        sq_value = t_main_diag[i] - chol_off_diag[i - 1] ** 2
-        chol_main_diag[i].copy_(torch.sqrt(sq_value))
-
-    res = torch.zeros_like(trid)
-    main_flattened_indices = batch_index * (batch_size * diag_size) + diag_index * (diag_size + 1)
-    off_flattened_indices = sum(
-        [off_batch_index * (batch_size * (diag_size - 1)), (off_diag_index + 1) * diag_size, off_diag_index]
-    )
-    res.view(-1).index_copy_(0, main_flattened_indices, chol_main_diag.view(-1))
-    res.view(-1).index_copy_(0, off_flattened_indices, chol_off_diag.view(-1))
-
-    if upper:
-        res = res.transpose(-1, -2)
-    return res
-
-
-def tridiag_batch_potrs(tensor, chol_trid, upper=True):
-    """
-    """
-    if not torch.is_tensor(chol_trid):
-        raise RuntimeError("tridiag_batch_potrf is only defined for tensors")
-
-    if not tensor.ndimension() == 3:
-        raise RuntimeError("Tensor should be 3 dimensional")
-
-    batch_size, diag_size, _ = chol_trid.size()
-    batch_index = torch.arange(0, batch_size, dtype=torch.long, device=tensor.device)
-    off_batch_index = batch_index.unsqueeze(1).repeat(1, diag_size - 1).view(-1)
-    batch_index = batch_index.unsqueeze(1).repeat(1, diag_size).view(-1)
-    diag_index = torch.arange(0, diag_size, dtype=torch.long, device=tensor.device)
-    diag_index = diag_index.unsqueeze(1).repeat(batch_size, 1).view(-1)
-    off_diag_index = torch.arange(0, diag_size - 1, dtype=torch.long, device=tensor.device)
-    off_diag_index = off_diag_index.unsqueeze(1).repeat(batch_size, 1).view(-1)
-
-    if upper:
-        chol_trid = chol_trid.transpose(-1, -2)
-
-    chol_main_diag = chol_trid[batch_index, diag_index, diag_index].view(batch_size, diag_size)
-    chol_off_diag = chol_trid[off_batch_index, off_diag_index + 1, off_diag_index].view(batch_size, diag_size - 1)
-
-    chol_solution = torch.empty_like(tensor)
-    chol_solution[:, 0, :].copy_(tensor[:, 0, :] / chol_main_diag[:, 0].unsqueeze(-1))
-    for i in range(1, diag_size):
-        inner_part = tensor[:, i, :]
-        inner_part = inner_part - chol_off_diag[:, i - 1].unsqueeze(-1) * chol_solution[:, i - 1, :]
-        chol_solution[:, i, :].copy_(inner_part / chol_main_diag[:, i].unsqueeze(-1))
-
-    solution = torch.empty_like(chol_solution)
-    solution[:, -1, :].copy_(chol_solution[:, -1, :] / chol_main_diag[:, -1].unsqueeze(-1))
-    for i in range(diag_size - 2, -1, -1):
-        inner_part = chol_solution[:, i, :] - chol_off_diag[:, i].unsqueeze(-1) * solution[:, i + 1, :]
-        solution[:, i, :].copy_(inner_part / chol_main_diag[:, i].unsqueeze(-1))
-
-    return solution
-
-
 __all__ = [
-    "Interpolation",
     "linear_cg",
     "StochasticLQ",
-    "left_interp",
-    "reverse",
-    "rcumsum",
-    "approx_equal",
-    "bdsmm",
-    "batch_potrf",
-    "batch_potrs",
-    "batch_symeig",
+    "cholesky",
+    "eig",
+    "fft",
     "grid",
+    "interpolation",
     "lanczos",
     "pivoted_cholesky",
-    "scale_to_bounds",
     "sparse",
-    "sparse_eye",
-    "sparse_getitem",
-    "sparse_repeat",
-    "to_sparse",
-    "tridiag_batch_potrf",
-    "tridiag_batch_potrs",
 ]
