+++ gpytorch/utils/interpolation.py
@@ -3,6 +3,8 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+from copy import deepcopy
+
 import torch
 
 
@@ -148,3 +150,109 @@ class Interpolation(object):
             interp_values = interp_values.mul(dim_interp_values.view(num_target_points, -1))
 
         return interp_indices, interp_values
+
+
+def left_interp(interp_indices, interp_values, rhs):
+    """
+    """
+    is_vector = rhs.ndimension() == 1
+
+    if is_vector:
+        res = rhs.index_select(0, interp_indices.view(-1)).view(*interp_values.size())
+        res = res.mul(interp_values)
+        res = res.sum(-1)
+        return res
+
+    else:
+        # Special cuda version -- this is faster on the GPU for some reason
+        if interp_indices.is_cuda:
+            if interp_indices.ndimension() == 3:
+                n_batch, n_data, n_interp = interp_indices.size()
+                interp_indices = interp_indices.contiguous().view(-1)
+                interp_values = interp_values.contiguous().view(-1, 1)
+
+                if rhs.ndimension() == 3:
+                    if rhs.size(0) == 1 and interp_indices.size(0) > 1:
+                        rhs = rhs.expand(interp_indices.size(0), rhs.size(1), rhs.size(2))
+                    batch_indices = torch.arange(0, n_batch, dtype=torch.long, device=rhs.device).unsqueeze_(1)
+                    batch_indices = batch_indices.repeat(1, n_data * n_interp).view(-1)
+                    res = rhs[batch_indices, interp_indices, :] * interp_values
+                else:
+                    res = rhs[interp_indices, :].unsqueeze(0) * interp_values
+                res = res.view(n_batch, n_data, n_interp, -1)
+                res = res.sum(-2)
+                return res
+            else:
+                n_data, n_interp = interp_indices.size()
+                interp_indices = interp_indices.contiguous().view(-1)
+                interp_values = interp_values.contiguous().view(-1, 1)
+                if rhs.ndimension() == 3:
+                    n_batch, _, n_cols = rhs.size()
+                    rhs = rhs.transpose(0, 1).contiguous().view(-1, n_batch * n_cols)
+                    res = rhs[interp_indices, :] * interp_values
+                    res = res.view(n_data, n_interp, n_batch, n_cols)
+                    res = res.sum(-2).transpose(0, 1).contiguous()
+                else:
+                    res = rhs[interp_indices, :] * interp_values
+                    res = res.view(n_data, n_interp, -1)
+                    res = res.sum(-2)
+                return res
+
+        # Special non-cuda version -- this is faster on the CPU
+        else:
+            interp_size = list(interp_indices.size()) + [rhs.size(-1)]
+            rhs_size = deepcopy(interp_size)
+            rhs_size[-3] = rhs.size()[-2]
+            interp_indices_expanded = interp_indices.unsqueeze(-1).expand(*interp_size)
+            res = rhs.unsqueeze(-2).expand(*rhs_size).gather(-3, interp_indices_expanded)
+            res = res.mul(interp_values.unsqueeze(-1).expand(interp_size))
+            return res.sum(-2)
+
+
+def left_t_interp(interp_indices, interp_values, rhs, output_dim):
+    """
+    """
+    from .. import dsmm
+
+    is_vector = rhs.ndimension() == 1
+    if is_vector:
+        rhs = rhs.unsqueeze(-1)
+
+    is_batch = rhs.ndimension() == 3
+    if not is_batch:
+        rhs = rhs.unsqueeze(0)
+    if not interp_indices.ndimension() == 3:
+        interp_indices = interp_indices.unsqueeze(0)
+        interp_values = interp_values.unsqueeze(0)
+
+    batch_size, n_data, n_interp = interp_values.size()
+    _, _, n_cols = rhs.size()
+
+    values = (rhs.unsqueeze(-2) * interp_values.unsqueeze(-1)).view(batch_size, n_data * n_interp, n_cols)
+
+    flat_interp_indices = interp_indices.contiguous().view(1, -1)
+    batch_indices = torch.arange(0, batch_size, dtype=torch.long, device=values.device).unsqueeze_(1)
+    batch_indices = batch_indices.repeat(1, n_data * n_interp).view(1, -1)
+    column_indices = torch.arange(0, n_data * n_interp, dtype=torch.long, device=values.device).unsqueeze_(1)
+    column_indices = column_indices.repeat(batch_size, 1).view(1, -1)
+
+    summing_matrix_indices = torch.cat([batch_indices, flat_interp_indices, column_indices])
+    summing_matrix_values = torch.ones(
+        batch_size * n_data * n_interp, dtype=interp_values.dtype, device=interp_values.device
+    )
+    size = torch.Size((batch_size, output_dim, n_data * n_interp))
+
+    type_name = summing_matrix_values.type().split(".")[-1]  # e.g. FloatTensor
+    if interp_values.is_cuda:
+        cls = getattr(torch.cuda.sparse, type_name)
+    else:
+        cls = getattr(torch.sparse, type_name)
+    summing_matrix = cls(summing_matrix_indices, summing_matrix_values, size)
+
+    res = dsmm(summing_matrix, values)
+
+    if not is_batch:
+        res = res.squeeze(0)
+    if is_vector:
+        res = res.squeeze(-1)
+    return res
