+++ gpytorch/utils/sparse.py
@@ -3,6 +3,8 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+from operator import mul
+
 import torch
 
 
@@ -78,4 +80,165 @@ def make_sparse_from_indices_and_values(interp_indices, interp_values, n_rows):
     res = cls(index_tensor, value_tensor, interp_size)
 
     # Wrap things as a variable, if necessary
+    return res
+
+
+def bdsmm(sparse, dense):
+    """
+    Batch dense-sparse matrix multiply
+    """
+    if sparse.ndimension() > 2:
+        batch_size, n_rows, n_cols = sparse.size()
+        batch_assignment = sparse._indices()[0]
+        indices = sparse._indices()[1:].clone()
+        indices[0].add_(n_rows, batch_assignment)
+        indices[1].add_(n_cols, batch_assignment)
+        sparse_2d = torch.sparse_coo_tensor(
+            indices,
+            sparse._values(),
+            torch.Size((batch_size * n_rows, batch_size * n_cols)),
+            dtype=sparse._values().dtype,
+            device=sparse._values().device,
+        )
+
+        if dense.size(0) == 1:
+            dense = dense.repeat(batch_size, 1, 1)
+        dense_2d = dense.contiguous().view(batch_size * n_cols, -1)
+        res = torch.dsmm(sparse_2d, dense_2d)
+        res = res.view(batch_size, n_rows, -1)
+        return res
+    elif dense.ndimension() == 3:
+        batch_size, _, n_cols = dense.size()
+        res = torch.dsmm(sparse, dense.transpose(0, 1).contiguous().view(-1, batch_size * n_cols))
+        res = res.view(-1, batch_size, n_cols)
+        res = res.transpose(0, 1).contiguous()
+        return res
+    else:
+        return torch.dsmm(sparse, dense)
+
+
+def sparse_eye(size):
+    """
+    Returns the identity matrix as a sparse matrix
+    """
+    indices = torch.arange(0, size).long().unsqueeze(0).expand(2, size)
+    values = torch.tensor(1.).expand(size)
+    cls = getattr(torch.sparse, values.type().split(".")[-1])
+    return cls(indices, values, torch.Size([size, size]))
+
+
+def sparse_getitem(sparse, idxs):
+    """
+    """
+    if not isinstance(idxs, tuple):
+        idxs = (idxs,)
+
+    if not sparse.ndimension() <= 2:
+        raise RuntimeError("Must be a 1d or 2d sparse tensor")
+
+    if len(idxs) > sparse.ndimension():
+        raise RuntimeError("Invalid index for %d-order tensor" % sparse.ndimension())
+
+    indices = sparse._indices()
+    values = sparse._values()
+    size = list(sparse.size())
+
+    for i, idx in list(enumerate(idxs))[::-1]:
+        if isinstance(idx, int):
+            del size[i]
+            mask = indices[i].eq(idx)
+            if sum(mask):
+                new_indices = torch.zeros(indices.size(0) - 1, sum(mask), dtype=indices.dtype, device=indices.device)
+                for j in range(indices.size(0)):
+                    if i > j:
+                        new_indices[j].copy_(indices[j][mask])
+                    elif i < j:
+                        new_indices[j - 1].copy_(indices[j][mask])
+                indices = new_indices
+                values = values[mask]
+            else:
+                indices.resize_(indices.size(0) - 1, 1).zero_()
+                values.resize_(1).zero_()
+
+            if not len(size):
+                return sum(values)
+
+        elif isinstance(idx, slice):
+            start, stop, step = idx.indices(size[i])
+            size = list(size[:i]) + [stop - start] + list(size[i + 1 :])
+            if step != 1:
+                raise RuntimeError("Slicing with step is not supported")
+            mask = indices[i].lt(stop) * indices[i].ge(start)
+            if sum(mask):
+                new_indices = torch.zeros(indices.size(0), sum(mask), dtype=indices.dtype, device=indices.device)
+                for j in range(indices.size(0)):
+                    new_indices[j].copy_(indices[j][mask])
+                new_indices[i].sub_(start)
+                indices = new_indices
+                values = values[mask]
+            else:
+                indices.resize_(indices.size(0), 1).zero_()
+                values.resize_(1).zero_()
+
+        else:
+            raise RuntimeError("Unknown index type")
+
+    return torch.sparse_coo_tensor(indices, values, torch.Size(size), dtype=values.dtype, device=values.device)
+
+
+def sparse_repeat(sparse, *repeat_sizes):
+    """
+    """
+    orig_ndim = sparse.ndimension()
+    new_ndim = len(repeat_sizes)
+    orig_nvalues = sparse._indices().size(1)
+
+    # Expand the number of dimensions to match repeat_sizes
+    indices = torch.cat(
+        [
+            torch.zeros(new_ndim - orig_ndim, orig_nvalues, dtype=torch.long, device=sparse._indices().device),
+            sparse._indices(),
+        ]
+    )
+    values = sparse._values()
+    size = [1] * (new_ndim - orig_ndim) + list(sparse.size())
+
+    # Expand each dimension
+    new_indices = torch.zeros(
+        indices.size(0), indices.size(1) * mul(*repeat_sizes), dtype=indices.dtype, device=indices.device
+    )
+    new_values = values.repeat(mul(*repeat_sizes))
+    new_size = [dim_size * repeat_size for dim_size, repeat_size in zip(size, repeat_sizes)]
+
+    # Fill in new indices
+    new_indices[:, :orig_nvalues].copy_(indices)
+    unit_size = orig_nvalues
+    for i in range(new_ndim)[::-1]:
+        repeat_size = repeat_sizes[i]
+        for j in range(1, repeat_size):
+            new_indices[:, unit_size * j : unit_size * (j + 1)].copy_(new_indices[:, :unit_size])
+            new_indices[i, unit_size * j : unit_size * (j + 1)] += j * size[i]
+        unit_size *= repeat_size
+
+    return torch.sparse_coo_tensor(
+        new_indices, new_values, torch.Size(new_size), dtype=new_values.dtype, device=new_values.device
+    )
+
+
+def to_sparse(dense):
+    """
+    """
+    mask = dense.ne(0)
+    indices = mask.nonzero()
+    if indices.storage():
+        values = dense[mask]
+    else:
+        indices = indices.resize_(1, dense.ndimension()).zero_()
+        values = torch.tensor(0, dtype=dense.dtype, device=dense.device)
+
+    # Construct sparse tensor
+    klass = getattr(torch.sparse, dense.type().split(".")[-1])
+    res = klass(indices.t(), values, dense.size())
+    if dense.is_cuda:
+        res = res.cuda()
     return res
