+++ test/functions/test_inv_quad_log_det.py
@@ -8,7 +8,7 @@ import torch
 import unittest
 import gpytorch
 from gpytorch.lazy import NonLazyTensor
-from gpytorch.utils import approx_equal
+from test._utils import approx_equal
 
 
 class TestInvQuadLogDetNonBatch(unittest.TestCase):
