+++ test/functions/test_root_decomposition.py
@@ -7,7 +7,7 @@ import os
 import torch
 import unittest
 from gpytorch.lazy import NonLazyTensor
-from gpytorch.utils import approx_equal
+from test._utils import approx_equal
 
 
 class TestRootDecomposition(unittest.TestCase):
