+++ test/lazy/test_chol_lazy_tensor.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals
 import torch
 import unittest
 from gpytorch.lazy import CholLazyTensor
-from gpytorch.utils import approx_equal
+from test._utils import approx_equal
 
 
 class TestCholLazyTensor(unittest.TestCase):
