+++ test/lazy/test_interpolated_lazy_tensor.py
@@ -9,7 +9,7 @@ import unittest
 import os
 import random
 from gpytorch.lazy import NonLazyTensor, InterpolatedLazyTensor
-from gpytorch.utils import approx_equal
+from test._utils import approx_equal
 
 
 class TestInterpolatedLazyTensor(unittest.TestCase):
