+++ test/lazy/test_sum_batch_lazy_tensor.py
@@ -9,7 +9,7 @@ import torch
 import unittest
 import gpytorch
 from gpytorch.lazy import SumBatchLazyTensor, NonLazyTensor
-from gpytorch.utils import approx_equal
+from test._utils import approx_equal
 
 
 class TestSumBatchLazyTensor(unittest.TestCase):
