+++ test/lazy/test_zero_lazy_tensor.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 
 import torch
 import unittest
-from gpytorch.utils import approx_equal
+from test._utils import approx_equal
 from gpytorch.lazy import ZeroLazyTensor
