+++ test/utils/test_lanczos.py
@@ -5,7 +5,7 @@ from __future__ import unicode_literals
 
 import torch
 import unittest
-from gpytorch.utils import approx_equal
+from test._utils import approx_equal
 from gpytorch.utils.lanczos import lanczos_tridiag
